Moved the Agenda Layout to Compose MP
