+++ androidApp/src/main/java/fr/paug/androidmakers/di/ViewModelModule.kt
@@ -9,7 +9,7 @@ import org.koin.dsl.module
 
 val androidViewModelModule = module {
   viewModel { MainActivityViewModel(get(), get(), get(), get(), get(), get(), get(), get()) }
-  factory { AgendaLayoutViewModel(get()) }
+
   factory { (sessionId: String) -> SessionDetailViewModel(sessionId, get(), get(), get(), get(), get()) }
-  factory { AgendaPagerViewModel(get(), get(), get()) }
+
 }
