--- iosApp/RobotConf/Model/Data/RoomData.swift
@@ -1,10 +1,0 @@
-//
-//  Copyright © 2020 Paris Android User Group. All rights reserved.
-//
-
-import Foundation
-
-struct RoomData: Decodable {
-    let roomId: String
-    let roomName: String
-}
