--- iosApp/RobotConf/Model/Room.swift
@@ -1,14 +1,0 @@
-//
-//  Room.swift
-//  RobotConf
-//
-//  Created by Antoine Robiez on 29/02/2024.
-//  Copyright © 2024 Djavan Bertrand. All rights reserved.
-//
-import Foundation
-
-struct Room: Hashable {
-    let uid: String
-    let name: String
-    let index: Int
-}
