--- iosApp/RobotConf/Model/Speaker.swift
@@ -1,15 +1,0 @@
-//
-//  Speaker.swift
-//  RobotConf
-//
-//  Created by Antoine Robiez on 29/02/2024.
-//  Copyright © 2024 Djavan Bertrand. All rights reserved.
-//
-import Foundation
-
-struct Speaker: Hashable {
-    let name: String
-    let photoUrl: URL
-    let company: String
-    let description: String
-}
