+++ shared/domain/src/iosMain/kotlin/fr/androidmakers/domain/utils/TextUtils.ios.kt
@@ -7,20 +7,27 @@ import kotlinx.cinterop.ExperimentalForeignApi
 import platform.Foundation.NSAttributedString
 import platform.Foundation.NSString
 import platform.Foundation.NSUTF8StringEncoding
+import platform.Foundation.create
 import platform.Foundation.dataUsingEncoding
 import platform.UIKit.NSDocumentTypeDocumentAttribute
+import platform.UIKit.NSDocumentTypeDocumentOption
 import platform.UIKit.NSHTMLTextDocumentType
 import platform.UIKit.create
 
-@OptIn(ExperimentalForeignApi::class, BetaInteropApi::class)
+@OptIn(ExperimentalForeignApi::class)
 actual fun String.removeHtmlTags(): String {
-  return (this as NSString).dataUsingEncoding(NSUTF8StringEncoding)?.let { data ->
-    val attributed = NSAttributedString.create(
-        data,
-        mapOf(NSDocumentTypeDocumentAttribute to NSHTMLTextDocumentType),
-        null,
-        null
-    )
-    return attributed?.string ?: ""
-  } ?: ""
+
+  if (this.isBlank()) {
+    return ""
+  } else {
+    return NSString.create(string = this).dataUsingEncoding(NSUTF8StringEncoding)?.let { data ->
+      val attributed = NSAttributedString.create(
+          data,
+          mapOf(NSDocumentTypeDocumentOption to NSHTMLTextDocumentType),
+          null,
+          null
+      )
+      return attributed?.string ?: ""
+    } ?: ""
+  }
 }
