+++ shared/ui/src/commonMain/kotlin/com/androidmakers/di/ViewModelModule.kt
@@ -1,5 +1,7 @@
 package com.androidmakers.di
 
+import com.androidmakers.ui.agenda.AgendaLayoutViewModel
+import com.androidmakers.ui.agenda.AgendaPagerViewModel
 import com.androidmakers.ui.speakers.SpeakerDetailsViewModel
 import com.androidmakers.ui.speakers.SpeakerListViewModel
 import com.androidmakers.ui.sponsors.SponsorsViewModel
@@ -11,5 +13,6 @@ val viewModelModule = module {
   factory { SponsorsViewModel(get()) }
   factory { VenueViewModel(get(), get(), get()) }
   factory { (speakerId: String) -> SpeakerDetailsViewModel(speakerId, get()) }
-
+  factory { AgendaLayoutViewModel(get()) }
+  factory { AgendaPagerViewModel(get(), get(), get()) }
 }
