+++ shared/ui/src/iosMain/kotlin/com/androidmakers/ui/MainLayoutViewController.kt
@@ -1,12 +1,8 @@
 package com.androidmakers.ui
 
-import androidx.compose.material3.DrawerValue
-import androidx.compose.material3.Text
-import androidx.compose.material3.rememberDrawerState
 import androidx.compose.ui.window.ComposeUIViewController
 import com.androidmakers.ui.about.AboutActions
 import com.androidmakers.ui.theme.AndroidMakersTheme
-import moe.tlaster.precompose.PreComposeApp
 import platform.UIKit.UIViewController
 
 
@@ -14,10 +10,6 @@ fun MainLayoutViewController(): UIViewController =
     ComposeUIViewController {
       AndroidMakersTheme {
         MainLayout(
-            agendaLayout = @androidx.compose.runtime.Composable {
-              val agendaFilterDrawerState = rememberDrawerState(DrawerValue.Closed)
-              Text("PLOP")
-            },
             aboutActions = AboutActions(
                 onFaqClick = {},
                 onCodeOfConductClick =  {},
